Cleans up some issues with specs + links all precompile implementations (#52)
I'm thinking a lot about the hackathon feedback and we've heard a few
times that the "flow" of the docs - esp. the specs - is a bit
unintuitive.

I think this is because we're still balancing with the docs not yet
ready for public release.

So, this PR just cleans up some obvious errors I found, and completes
the precompiles page etc.

In the next big update, we can think about whether (and how) it makes
sense to move "Example Flows" into the docs so people find them earlier,
in tandem with pages demonstrating working code.